+++ corpus/.clj-kondo/hooks/new.clj
@@ -5,3 +5,6 @@
   (api/reg-finding! (assoc (meta node)
                            :message (str "Interop is no good! " (api/generated-node? (first (:children node))))
                            :type :interop)))
+
+(defmacro new-macroexpand [& body]
+  `(do ~body))